bump deps
standard abi encode
